switch back to float

+++ components/asic/bm1366.c
@@ -138,7 +138,7 @@ void BM1366_set_version_mask(uint32_t version_mask)
     _send_BM1366(TYPE_CMD | GROUP_ALL | CMD_WRITE, version_cmd, 6, BM1366_SERIALTX_DEBUG);
 }
 
-void BM1366_send_hash_frequency(double target_freq)
+void BM1366_send_hash_frequency(float target_freq)
 {
     // default 200Mhz if it fails
     unsigned char freqbuf[9] = {0x00, 0x08, 0x40, 0xA0, 0x02, 0x41}; // freqbuf - pll0_parameter
@@ -195,13 +195,13 @@ void BM1366_send_hash_frequency(double target_freq)
     ESP_LOGI(TAG, "Setting Frequency to %.2fMHz (%.2f)", target_freq, newf);
 }
 
-static void do_frequency_ramp_up(double target_frequency) {
+static void do_frequency_ramp_up(float target_frequency) {
     ESP_LOGI(TAG, "Ramping up frequency from %.2f MHz to %.2f MHz", current_frequency, target_frequency);
     do_frequency_transition(target_frequency, BM1366_send_hash_frequency, 1366);
 }
 
 // Add a public function for external use
-bool BM1366_set_frequency(double target_freq) {
+bool BM1366_set_frequency(float target_freq) {
     return do_frequency_transition(target_freq, BM1366_send_hash_frequency, 1366);
 }
